+++ packages/local_auth/local_auth_android/lib/local_auth_android.dart
@@ -49,28 +49,24 @@ class LocalAuthAndroid extends LocalAuthPlatform {
 
   @override
   Future<bool> deviceSupportsBiometrics() async {
-    return (await getEnrolledBiometrics()).isNotEmpty;
+    return (await _channel.invokeMethod<bool>('deviceSupportsBiometrics')) ??
+        false;
   }
 
   @override
   Future<List<BiometricType>> getEnrolledBiometrics() async {
     final List<String> result = (await _channel.invokeListMethod<String>(
-          'getAvailableBiometrics',
+          'getEnrolledBiometrics',
         )) ??
         <String>[];
     final List<BiometricType> biometrics = <BiometricType>[];
     for (final String value in result) {
       switch (value) {
-        case 'face':
-          biometrics.add(BiometricType.face);
+        case 'weak':
+          biometrics.add(BiometricType.weak);
           break;
-        case 'fingerprint':
-          biometrics.add(BiometricType.fingerprint);
-          break;
-        case 'iris':
-          biometrics.add(BiometricType.iris);
-          break;
-        case 'undefined':
+        case 'strong':
+          biometrics.add(BiometricType.strong);
           break;
       }
     }