+++ script/tool/CHANGELOG.md
@@ -1,10 +1,11 @@
-## NEXT
+## 0.8.2+1
 
 - Adds a new `readme-check` command.
 - Updates `publish-plugin` command documentation.
 - Fixes `all-plugins-app` to preserve the original application's Dart SDK
   version to avoid changing language feature opt-ins that the template may
   rely on.
+- Fixes `custom-test` to run `pub get` before running Dart test scripts.
 
 ## 0.8.2